better configuration typing
