+++ scripts/rules/types/auth0-rules.d.ts
@@ -2,9 +2,14 @@
 export {}
 
 declare global {
-  namespace NodeJS {
-    interface Global {
-      CACHE_KEY: string
-    }
+  /** Extend this interface, it will appear as the type of the global configuration object */
+  interface IAuth0RuleConfiguration {
+    POSTGRES_USERNAME: string
+    POSTGRES_PASSWORD: string
+    POSTGRES_HOST: string
+    POSTGRES_DATABASE: string
+    POSTGRES_PORT?: string
+    AUTH0_CLIENT_ID: string
+    AUTH0_CLIENT_SECRET: string
   }
 }
